Folders refactored
