+++ src/services/calculateAmount.js
@@ -1,8 +1,7 @@
-const output = require('./output')
+const output = require('../utils/output')
 module.exports = (resultArray) => {
   let amount = 1
   // TODO output will declare with a global variable
-  console.log('Result:')
   resultArray.forEach((value, index) => {
     if (resultArray[index + 1]) {
       if (value === resultArray[index + 1]) {
