+++ src/utils/checkUserInput.js
